wip: moving more into the backend
