+++ yew/src/utils.rs
@@ -1,83 +1,6 @@
 //! This module contains useful utilities to get information about the current document.
-
-use anyhow::{anyhow, Error};
-use cfg_if::cfg_if;
-use cfg_match::cfg_match;
 use std::marker::PhantomData;
 use yew::html::ChildrenRenderer;
-cfg_if! {
-    if #[cfg(feature = "std_web")] {
-        use stdweb::web::{Document, Window};
-    } else if #[cfg(feature = "web_sys")] {
-        use web_sys::{Document, Window};
-    } else if #[cfg(feature = "static_render")] {
-        use crate::smr::mock::{Document, Window, get_window, get_document};
-    }
-}
-
-/// Returns the current window. This function will panic if there is no available window.
-pub fn window() -> Window {
-    cfg_match! {
-        feature = "std_web" => stdweb::web::window(),
-        feature = "web_sys" => web_sys::window().expect("no window available"),
-        feature = "static_render" => get_window(),
-    }
-}
-
-/// Returns the current document.
-pub fn document() -> Document {
-    cfg_match! {
-        feature = "std_web" => stdweb::web::document(),
-        feature = "web_sys" => window().document().unwrap(),
-        feature = "static_render" => get_document(),
-    }
-}
-
-/// Returns the `host` for the current document. Useful for connecting to the server which serves
-/// the app.
-pub fn host() -> Result<String, Error> {
-    let location = document()
-        .location()
-        .ok_or_else(|| anyhow!("can't get location"))?;
-
-    #[cfg(feature = "std_web")]
-    let host = location.host().map_err(Error::from)?;
-
-    #[cfg(feature = "web_sys")]
-    let host = location.host().map_err(|e| {
-        anyhow!(e
-            .as_string()
-            .unwrap_or_else(|| String::from("error not recoverable")),)
-    })?;
-
-    #[cfg(feature = "static_render")]
-    let host = crate::smr::mock::get_host();
-
-    Ok(host)
-}
-
-/// Returns the `origin` of the current window.
-pub fn origin() -> Result<String, Error> {
-    let location = window().location();
-
-    #[cfg(feature = "std_web")]
-    let location = location.ok_or_else(|| anyhow!("can't get location"))?;
-
-    #[cfg(feature = "std_web")]
-    let origin = location.origin().map_err(Error::from)?;
-
-    #[cfg(feature = "web_sys")]
-    let origin = location.origin().map_err(|e| {
-        anyhow!(e
-            .as_string()
-            .unwrap_or_else(|| String::from("error not recoverable")),)
-    })?;
-
-    #[cfg(feature = "static_render")]
-    let origin = crate::smr::mock::get_origin();
-
-    Ok(origin)
-}
 
 /// Map IntoIterator<Item=Into<T>> to Iterator<Item=T>
 pub fn into_node_iter<IT, T, R>(it: IT) -> impl Iterator<Item = R>
